requirements.txt
paho-mqtt==2.0.0
pycryptodome==3.20.0
pyserial==3.5
PyYAML==6.0.1